refactor number builder directory structure
