+++ src/de/access.rs
@@ -1,4 +1,5 @@
 pub mod r#enum;
 pub mod jsonc;
 pub mod map;
+pub mod number;
 pub mod seq;
